refactor docstring parse
